Marvericks Support
- Add InstallESD version auto detecting

+++ InstallESD.dmg.kext.tool
@@ -2,38 +2,33 @@
 
 InstallESD_dmg_kext_tool () {
   Help=$(cat <<EOF
-usage: $0 [-BIL] [-i InstallESD.dmg] [-o Output.dmg] [--] [kext ...]
+usage: $0 [-b] [-v X.Y] [-i InstallESD.dmg] [-o Output.dmg] [--] [kext ...]
        $0 [-h]
 
 OPTIONS:
   -h  Print Help (this message) and exit
+  -b  Use BaseSystem as container
   -i  Location of InstallESD.dmg
   -o  Location of output
-  -B  BaseSystem mode (default)
-  -I  InstallESD mode
-  -L  Legacy mode for Mac OS X version 10.6 (Snow Leopard) and earlier
+  -v  Force X.Y as InstallESD version
 
 EXAMPLE:
   $0 -i InstallESD.dmg -o Output.dmg -- NullCPUPowerManagement.kext
 
 EOF)
-  Mode=B
-  while getopts hi:o:IBL opt; do
+  while getopts hbi:o:v: opt; do
     case $opt in
+      b)
+        BaseSystem=1
+        ;;
       i)
-        Input=$OPTARG
+        Input_DMG=$OPTARG
         ;;
       o)
-        Output=$OPTARG
+        Output_DMG=$OPTARG
         ;;
-      I)
-        Mode=I
-        ;;
-      B)
-        Mode=B
-        ;;
-      L)
-        Mode=L
+      v)
+        Version=$OPTARG
         ;;
       h)
         echo "$Help"
@@ -47,92 +42,149 @@ EOF)
   done
 
   shift $(($OPTIND - 1))
-  if [ -z "$Input" ] || [ -z "$Output" ]; then
+  if [ -z "$Input_DMG" ] || [ -z "$Output_DMG" ]; then
     echo "$Help" >&2
     return 1
   fi
-  if [ ! -f "$Input" ]; then
+  if [ ! -f "$Input_DMG" ]; then
     echo "InstallESD.dmg not found." >&2
     return 1
   fi
-  if [ ! "$(echo "${Input##*.}" | tr "[:upper:]" "[:lower:]")" = "dmg" ]; then
+  if [ ! "$(echo "${Input_DMG##*.}" | tr "[:upper:]" "[:lower:]")" = "dmg" ]; then
     echo "Only dmg format is supported." >&2
     return 1
   fi
-  if [ -e "$Output" ]; then
-    echo "$Output already exists." >&2
+  if [ -e "$Output_DMG" ]; then
+    echo "$Output_DMG already exists." >&2
     return 1
   fi
-  if [ ! "$Mode" = "B" ] && [ "$#" -eq 0 ]; then
-    echo "This mode requires at least one kext." >&2
+  if [ -z "$BaseSystem" ] && [ "$#" -eq 0 ]; then
+    echo "Require at least one kext." >&2
     return 1
   fi
 
   Kexts=( "$@" )
-  echo "Checking Kexts"
-  for Kext in "${Kexts[@]}"; do
-    KextBaseName=$(basename -- "$Kext")
-    if [ -d "$Kext" ] && [ "${KextBaseName##*.}" = "kext" ] && [ -f "$Kext/Contents/MacOS/${KextBaseName%.*}" ]; then
-      echo "✓ $KextBaseName"
-    else
-      echo "✗ $Kext" >&2
-      return 1
-    fi
-  done
+  if [ "${#Kexts[@]}" -gt 0 ]; then
+    echo
+    echo "Checking Kexts"
+    for Kext in "${Kexts[@]}"; do
+      KextBaseName=$(basename -- "$Kext")
+      if [ -d "$Kext" ] && [ "${KextBaseName##*.}" = "kext" ] && [ -f "$Kext/Contents/MacOS/${KextBaseName%.*}" ]; then
+        echo "✓ $KextBaseName"
+      else
+        echo "✗ $Kext" >&2
+        return 1
+      fi
+    done
+  fi
 
   Temp=$(mktemp -d "/tmp/InstallESD.dmg.kext.tool.XXXXXXXX")
 
-  InstallESD_DMG=$Input
+  InstallESD_DMG=$Input_DMG
   InstallESD_DMG_Format=$(hdiutil imageinfo -format "$InstallESD_DMG")
   InstallESD=$Temp/InstallESD
   InstallESD_BaseSystem_DMG=$InstallESD/BaseSystem.dmg
-  [ "$Mode" = "L" ] &&
-    InstallESD_BaseSystem_DMG=$Input
+
   RW_BaseSystem_DMG=$Temp/RW_BaseSystem.dmg
   RW_BaseSystem=$Temp/RW_BaseSystem
-  RW_BaseSystem_kernelcache=$RW_BaseSystem/System/Library/Caches/com.apple.kext.caches/Startup/kernelcache
-  RW_BaseSystem_mkext2=$RW_BaseSystem/System/Library/Caches/com.apple.kext.caches/Startup/Extensions.mkext
-  RW_BaseSystem_mkext1=$RW_BaseSystem/System/Library/Extensions.mkext
   RW_InstallESD_DMG=$Temp/RW_InstallESD.dmg
   RW_InstallESD=$Temp/RW_InstallESD
   RW_InstallESD_BaseSystem_DMG=$RW_InstallESD/BaseSystem.dmg
-  RW_InstallESD_kernelcache=$RW_InstallESD/kernelcache
-  OSInstall_PKG=$RW_BaseSystem/System/Installation/Packages/OSInstall.pkg
-  Kernel=$RW_BaseSystem/mach_kernel
-  [ "$Mode" = "I" ] &&
-    OSInstall_PKG=$RW_InstallESD/Packages/OSInstall.pkg &&
-    Kernel=$RW_InstallESD/mach_kernel
-  OSInstall=$Temp/OSInstall
-  OSInstall_Script=$OSInstall/Scripts/postinstall_actions/installAdditionalKexts
 
-  if [ ! "$Mode" = "L" ]; then
+  BaseSystemBinaries_PKG=$InstallESD/Packages/BaseSystemBinaries.pkg
+  BaseSystemBinaries=$Temp/BaseSystemBinaries
+  OSInstall_PKG=$RW_InstallESD/Packages/OSInstall.pkg
+  OSInstall=$Temp/OSInstall
+  OSInstall_Script=$OSInstall/Scripts/postinstall_actions/kext.tool
+
+  Startup_kernelcache=/System/Library/Caches/com.apple.kext.caches/Startup/kernelcache
+  Startup_mkext2=/System/Library/Caches/com.apple.kext.caches/Startup/Extensions.mkext
+  Startup_mkext1=/System/Library/Extensions.mkext
+
+  if [ -z "$Version" ] || [ "$(echo "$Version >= 10.7" | bc)" -eq 1 ]; then
     echo
     echo "Mounting Install ESD"
     mkdir "$InstallESD"
     hdiutil attach -quiet -nobrowse -noverify -mountpoint "$InstallESD" "$InstallESD_DMG"
-    if [ -f "$InstallESD_BaseSystem_DMG" ]; then
-      InstallESD_BaseSystem_DMG_Format=$(hdiutil imageinfo -format "$InstallESD_BaseSystem_DMG")
-    else
+    if [ "$?" -ne 0 ]; then
       hdiutil detach -quiet "$InstallESD"
-      if [ "$?" -eq 0 ]; then
-        echo "BaseSystem.dmg not found in InstallESD.dmg." >&2
-      else
-        echo "Failed to mount InstallESD.dmg." >&2
-      fi
+      echo "Failed to mount InstallESD.dmg." >&2
       rm -rf "$Temp"
       return 1
     fi
+  fi
+
+  if [ -z "$Version" ]; then
+    echo
+    echo "Detecting Install ESD Version"
+    if [ -f "$InstallESD_BaseSystem_DMG" ]; then
+      if [ ! -f "$InstallESD/mach_kernel" ]; then
+        Version=10.9
+      else
+        Version=10.8
+        Version=10.7
+        # 10.7 and 10.8 shares same structure
+      fi
+    elif [ -f "$InstallESD$Startup_mkext2" ]; then
+      Version=10.6
+    elif [ -f "$InstallESD$Startup_mkext1" ]; then
+      Version=10.5
+    else
+      echo "Unknown InstallESD Version." >&2
+      return 1
+    fi
+    if [ "$(echo "$Version < 10.7" | bc)" -eq 1 ]; then
+      echo
+      echo "Unmounting Install ESD"
+      hdiutil detach -quiet "$InstallESD"
+      rm -r "$InstallESD"
+    fi
+  fi
+
+  if [ "$(echo "$Version >= 10.7" | bc)" -eq 1 ]; then
+    InstallESD_BaseSystem_DMG_Format=$(hdiutil imageinfo -format "$InstallESD_BaseSystem_DMG")
+  elif [ "$(echo "$Version >= 10.5" | bc)" -eq 1 ]; then
+    InstallESD_BaseSystem_DMG=$InstallESD_DMG
+    InstallESD_BaseSystem_DMG_Format=$InstallESD_DMG_Format
+  fi
+
+  if [ "$(echo "$Version >= 10.9" | bc)" -eq 1 ]; then
+    if [ "${#Kexts[@]}" -gt 0 ]; then
+      echo
+      echo "Extracting Kernel"
+      pkgutil --expand "$BaseSystemBinaries_PKG" "$BaseSystemBinaries"
+      case "$(file --brief --mime-type "$BaseSystemBinaries/Payload")" in
+        application/x-bzip2)
+          mv "$BaseSystemBinaries/Payload" "$BaseSystemBinaries/Payload.cpio.bz2"
+          bunzip2 "$BaseSystemBinaries/Payload.cpio.bz2"
+          ;;
+        application/x-gzip)
+          mv "$BaseSystemBinaries/Payload" "$BaseSystemBinaries/Payload.cpio.gz"
+          gunzip "$BaseSystemBinaries/Payload.cpio.gz"
+          ;;
+      esac
+      echo "/mach_kernel" | cpio -p -d -I "$BaseSystemBinaries/Payload.cpio" -- "$BaseSystemBinaries/Payload"
+    fi
+    Mach_Kernel=$BaseSystemBinaries/Payload/mach_kernel
+  elif [ "$(echo "$Version >= 10.7" | bc)" -eq 1 ]; then
+    Mach_Kernel=$RW_InstallESD/mach_kernel
+  elif [ "$(echo "$Version >= 10.5" | bc)" -eq 1 ]; then
+    Mach_Kernel=$RW_BaseSystem/mach_kernel
+  fi
+
+  if [ "$(echo "$Version <= 10.6" | bc)" -eq 1 ] || [ -n "$BaseSystem" ]; then
+    OSInstall_PKG=$RW_BaseSystem/System/Installation/Packages/OSInstall.pkg
+    Mach_Kernel=$RW_BaseSystem/mach_kernel
   fi
 
   echo
   echo "Creating Temporary Base System in UDRW format"
   hdiutil convert -format UDRW -o "$RW_BaseSystem_DMG" "$InstallESD_BaseSystem_DMG"
 
-  if [ "$Mode" = "B" ]; then
+  if [ -n "$BaseSystem" ]; then
     echo
-    RW_BaseSystem_Size_Sectors=$(( $(hdiutil resize -limits "$InstallESD_DMG" | tail -n 1 | cut -f 1) + $(hdiutil resize -limits "$InstallESD_BaseSystem_DMG" | tail -n 1 | cut -f 1) ))
-    echo "Resizing Temporary Base System to $RW_BaseSystem_Size_Sectors blocks"
-    hdiutil resize -sectors "$RW_BaseSystem_Size_Sectors" "$RW_BaseSystem_DMG"
+    echo "Resizing Temporary Base System"
+    hdiutil resize -sectors "$(( $(hdiutil resize -limits "$InstallESD_DMG" | tail -n 1 | cut -f 1) + $(hdiutil resize -limits "$InstallESD_BaseSystem_DMG" | tail -n 1 | cut -f 1) ))" "$RW_BaseSystem_DMG"
   fi
 
   echo
@@ -140,10 +192,12 @@ EOF)
   echo "Mounting Temporary Base System"
   hdiutil attach -owners on -nobrowse -mountpoint "$RW_BaseSystem" "$RW_BaseSystem_DMG"
 
-  if [ "$Mode" = "B" ]; then
-    echo
-    echo "Copying Kernel"
-    sudo -p "Please enter %u's password:" cp "$InstallESD/mach_kernel" "$RW_BaseSystem/mach_kernel"
+  if [ -n "$BaseSystem" ]; then
+    if [ "$(echo "$Version <= 10.8" | bc)" -eq 1 ] && [ "$(echo "$Version >= 10.7" | bc)" -eq 1 ]; then
+      echo
+      echo "Copying Kernel"
+      sudo -p "Please enter %u's password:" cp "$InstallESD/mach_kernel" "$RW_BaseSystem/mach_kernel"
+    fi
 
     echo
     echo "Copying Packages"
@@ -151,21 +205,14 @@ EOF)
     sudo -p "Please enter %u's password:" cp -R "$InstallESD/Packages" "$RW_BaseSystem/System/Installation/Packages"
   fi
 
-  if [ ! "$Mode" = "L" ]; then
+  if [ "$(echo "$Version >= 10.7" | bc)" -eq 1 ]; then
     echo
     echo "Unmounting Install ESD"
     hdiutil detach -quiet "$InstallESD"
     rm -r "$InstallESD"
   fi
 
-  echo
-  echo "Copying Kexts"
-  for Kext in "${Kexts[@]}"; do
-    KextBaseName=$(basename -- "$Kext")
-    sudo -p "Please enter %u's password:" cp -R "$Kext" "$RW_BaseSystem/System/Library/Extensions/$KextBaseName" && echo "✓ $KextBaseName"
-  done
-
-  if [ "$Mode" = "I" ]; then
+  if [ "$(echo "$Version >= 10.7" | bc)" -eq 1 ] && [ -z "$BaseSystem" ]; then
     echo
     echo "Creating Temporary Install ESD in UDRW format"
     hdiutil convert -format UDRW -ov -o "$RW_InstallESD_DMG" "$InstallESD_DMG"
@@ -176,63 +223,67 @@ EOF)
     hdiutil attach -owners on -nobrowse -mountpoint "$RW_InstallESD" "$RW_InstallESD_DMG"
   fi
 
-  if [ ! "$Mode" = "L" ]; then
+  if [ "${#Kexts[@]}" -gt 0 ]; then
     echo
-    echo "Rebuilding kernelcache"
-    sudo -p "Please enter %u's password:" kextcache -v 0 -prelinked-kernel "$RW_BaseSystem_kernelcache" -kernel "$Kernel" -volume-root "$RW_BaseSystem" -- "$RW_BaseSystem/System/Library/Extensions"
+    echo "Copying Kexts"
+    for Kext in "${Kexts[@]}"; do
+      KextBaseName=$(basename -- "$Kext")
+      sudo -p "Please enter %u's password:" cp -R "$Kext" "$RW_BaseSystem/System/Library/Extensions/$KextBaseName" && echo "✓ $KextBaseName"
+    done
   fi
 
-  if [ "$Mode" = "L" ]; then
+  if [ "$(echo "$Version >= 10.7" | bc)" -eq 1 ]; then
+    echo
+    echo "Rebuilding kernelcache"
+    sudo -p "Please enter %u's password:" kextcache -v 0 -prelinked-kernel "$RW_BaseSystem$Startup_kernelcache" -kernel "$Mach_Kernel" -volume-root "$RW_BaseSystem" -- "$RW_BaseSystem/System/Library/Extensions"
+  elif [ "$(echo "$Version >= 10.5" | bc)" -eq 1 ]; then
     echo
     echo "Rebuilding mkext cache"
-    if [ -f "$RW_BaseSystem_mkext2" ]; then
-      sudo -p "Please enter %u's password:" kextcache -v 0 -a i386 -a x86_64 -mkext "$RW_BaseSystem_mkext2" -kernel "$Kernel" -volume-root "$RW_BaseSystem" -- "$RW_BaseSystem/System/Library/Extensions"
-      [ -f "$RW_BaseSystem_mkext1" ] &&
-        sudo -p "Please enter %u's password:" cp "$RW_BaseSystem_mkext2" "$RW_BaseSystem_mkext1"
+    if [ "$(echo "$Version >= 10.6" | bc)" -eq 1 ]; then
+      sudo -p "Please enter %u's password:" kextcache -v 0 -a i386 -a x86_64 -mkext "$RW_BaseSystem$Startup_mkext2" -kernel "$Mach_Kernel" -volume-root "$RW_BaseSystem" -- "$RW_BaseSystem/System/Library/Extensions"
+      [ -f "$RW_BaseSystem$Startup_mkext1" ] &&
+        sudo -p "Please enter %u's password:" cp "$RW_BaseSystem$Startup_mkext2" "$RW_BaseSystem$Startup_mkext1"
     else
-      sudo -p "Please enter %u's password:" kextcache -v 0 -a ppc -a i386 -mkext "$RW_BaseSystem_mkext1" -kernel "$Kernel" -volume-root "$RW_BaseSystem" -- "$RW_BaseSystem/System/Library/Extensions"
+      sudo -p "Please enter %u's password:" kextcache -v 0 -a ppc -a i386 -mkext "$RW_BaseSystem$Startup_mkext1" -kernel "$Mach_Kernel" -volume-root "$RW_BaseSystem" -- "$RW_BaseSystem/System/Library/Extensions"
     fi
   fi
 
-  if [ "$Mode" = "I" ]; then
+
+  if [ "$(echo "$Version <= 10.8" | bc)" -eq 1 ] && [ "$(echo "$Version >= 10.7" | bc)" -eq 1 ] && [ -z "$BaseSystem" ]; then
     echo
     echo "Updating kernelcache on Temporary Install ESD"
-    sudo -p "Please enter %u's password:" cp "$RW_BaseSystem_kernelcache" "$RW_InstallESD_kernelcache"
+    sudo -p "Please enter %u's password:" cp "$RW_BaseSystem$Startup_kernelcache" "$RW_InstallESD/kernelcache"
+    sudo -p "Please enter %u's password:" chflags hidden "$RW_InstallESD/kernelcache"
   fi
 
-  echo
-  echo "Creating OSInstall Script for Kexts"
-  pkgutil --expand "$OSInstall_PKG" "$OSInstall"
-  touch "$OSInstall_Script" && chmod a+x "$OSInstall_Script"
-  echo "#!/bin/sh" > "$OSInstall_Script"
-  echo >> "$OSInstall_Script"
-  for Kext in "${Kexts[@]}"; do
-    KextBaseName=$(basename "$Kext")
-    echo "logger -p install.info \"Installing $KextBaseName\"" >> "$OSInstall_Script"
-    echo "/bin/cp -R \"/System/Library/Extensions/$KextBaseName\" \"\$3/System/Library/Extensions/$KextBaseName\"" >> "$OSInstall_Script"
+  if [ "${#Kexts[@]}" -gt 0 ]; then
+    echo
+    echo "Creating OSInstall Script for Kexts"
+    pkgutil --expand "$OSInstall_PKG" "$OSInstall"
+    touch "$OSInstall_Script" && chmod a+x "$OSInstall_Script"
+    echo "#!/bin/sh" > "$OSInstall_Script"
     echo >> "$OSInstall_Script"
-  done
-  echo "exit 0" >> "$OSInstall_Script"
-  sudo -p "Please enter %u's password:" pkgutil --flatten "$OSInstall" "$OSInstall_PKG"
-  rm -r "$OSInstall"
+    for Kext in "${Kexts[@]}"; do
+      KextBaseName=$(basename "$Kext")
+      echo "logger -p install.info \"Installing $KextBaseName\"" >> "$OSInstall_Script"
+      echo "/bin/cp -R \"/System/Library/Extensions/$KextBaseName\" \"\$3/System/Library/Extensions/$KextBaseName\"" >> "$OSInstall_Script"
+      echo >> "$OSInstall_Script"
+    done
+    echo "exit 0" >> "$OSInstall_Script"
+    sudo -p "Please enter %u's password:" pkgutil --flatten "$OSInstall" "$OSInstall_PKG"
+    rm -r "$OSInstall"
+  fi
 
   echo
   echo "Unmounting Temporary Base System"
   hdiutil detach -quiet "$RW_BaseSystem"
   rm -r "$RW_BaseSystem"
 
-  if [ "$Mode" = "B" ] || [ "$Mode" = "L" ]; then
-    echo
-    echo "Converting Temporary Base System to $InstallESD_DMG_Format format"
-    hdiutil convert -format "$InstallESD_DMG_Format" -o "$Output" "$RW_BaseSystem_DMG"
-    rm "$RW_BaseSystem_DMG"
-  fi
-
-  if [ "$Mode" = "I" ]; then
+  if [ "$(echo "$Version >= 10.7" | bc)" -eq 1 ] && [ -z "$BaseSystem" ]; then
     echo
     echo "Convert Temporary Base System to $InstallESD_BaseSystem_DMG_Format format"
     sudo -p "Please enter %u's password:" hdiutil convert -format "$InstallESD_BaseSystem_DMG_Format" -ov -o "$RW_InstallESD_BaseSystem_DMG" "$RW_BaseSystem_DMG"
-    sudo -p "Please enter %u's password:" chflags hidden "$RW_InstallESD_BaseSystem_DMG" "$RW_InstallESD_kernelcache"
+    sudo -p "Please enter %u's password:" chflags hidden "$RW_InstallESD_BaseSystem_DMG"
     rm "$RW_BaseSystem_DMG"
 
     echo
@@ -242,8 +293,13 @@ EOF)
 
     echo
     echo "Converting Temporary Install ESD to $InstallESD_DMG_Format format"
-    hdiutil convert -format "$InstallESD_DMG_Format" -o "$Output" "$RW_InstallESD_DMG"
+    hdiutil convert -format "$InstallESD_DMG_Format" -o "$Output_DMG" "$RW_InstallESD_DMG"
     rm "$RW_InstallESD_DMG"
+  elif [ "$(echo "$Version >= 10.5" | bc)" -eq 1 ]; then
+    echo
+    echo "Converting Temporary Base System to $InstallESD_BaseSystem_DMG_Format format"
+    hdiutil convert -format "$InstallESD_BaseSystem_DMG_Format" -o "$Output_DMG" "$RW_BaseSystem_DMG"
+    rm "$RW_BaseSystem_DMG"
   fi
 
   rm -rf "$Temp"
